Button and ammo
Button: when the user clicks on the component, it pauses or resumes the game. It changes the text to the action that takes place if the user clicks on the button (pause or resume)

Ammo: The user can fire 10 times before they run out of ammo, which they have to reload to be able to fire again. When the ammo count reaches zero, the user loses the ability to fire.